do not create pod if it is already running do not clean with a yaml file if file not exist

+++ automation/performance_operator_cleanup.sh
@@ -8,7 +8,7 @@ source ./functions.sh
 parse_args $@
 
 echo "Removing performance profile ..."
-oc delete -f ${MANIFEST_DIR}/performance_profile.yaml
+[ -e ${MANIFEST_DIR}/performance_profile.yaml ] && oc delete -f ${MANIFEST_DIR}/performance_profile.yaml
 
 if [[ "${WAIT_MCP}" == "true" ]]; then
     wait_mcp
